api/requirements.txt
flask
flask-cors
flask-sqlalchemy
marshmallow
psycopg2
pyjwt
python-dateutil